+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - control: Import MPC balancer class from the Pink balancer
 - cpp: Add documentation to the moteus actuation interface
 - cpp: Controller pipeline run by the spine after observers
+- envs: Abstract base class for environments named `UpkieEnv`
 - envs: Add Genesis backend
 - envs: Add PyBullet backend
 - envs: Add internal `upkie.envs.backends` submodule
@@ -51,7 +52,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - cpp: Limit rate of SynchronousClock skip warnings and report frequency drop
 - cpp: Update mpacklog.cpp to version 3.1.2
 - envs: Allow gain scales larger than one in servo environments
-- envs: Base class for environments is now `UpkieEnv`
 - envs: Class `UpkieServos` is now a subclass of `UpkieEnv`
 - envs: Drop Gymnasium environment versioning
 - envs: Remove the CPU bottleneck, allowing parallel simulations.